refactor: hook 위치 수정
